# Move CLI related commands to `nu-cli`