--- a/crates/nu-cli/src/commands/keybindings_listen.rs
+++ b/crates/nu-cli/src/commands/keybindings_listen.rs
@@ -144,14 +144,3 @@ fn print_events_helper(event: Event) -> Result<Value, ShellError> {
         Ok(record)
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use crate::KeybindingsListen;
-
-    #[test]
-    fn examples_work_as_expected() {
-        use crate::test_examples;
-        test_examples(KeybindingsListen {})
-    }
-}