--- a/crates/nu-cli/src/lib.rs
+++ b/crates/nu-cli/src/lib.rs
@@ -1,6 +1,7 @@
 mod commands;
 mod completions;
 mod config_files;
+mod eval_cmds;
 mod eval_file;
 mod menus;
 mod nu_highlight;
@@ -13,9 +14,10 @@ mod syntax_highlight;
 mod util;
 mod validation;
 
-pub use commands::evaluate_commands;
+pub use commands::add_cli_context;
 pub use completions::{FileCompletion, NuCompleter};
 pub use config_files::eval_config_contents;
+pub use eval_cmds::evaluate_commands;
 pub use eval_file::evaluate_file;
 pub use menus::{DescriptionMenu, NuHelpCompleter};
 pub use nu_command::util::get_init_cwd;