… Cargo.lock
… crates/nu-cli/src/commands/commandline.rs
… crates/nu-cli/src/commands/history.rs
… crates/nu-cli/src/commands/history_session.rs
… crates/nu-cli/src/commands/keybindings.rs
… crates/nu-cli/src/commands/keybindings_default.rs
… crates/nu-cli/src/commands/keybindings_list.rs
… crates/nu-cli/src/eval_cmds.rs