@stormasm 2023-04-10
@sholderbach this is a nice PR...

I have no idea why commandline was ever in nu-cmd-lang in the first place 😄 ?

cool so it cleans up nu-command as well

misc is a catch all bucket and having just tutor in there now seems good as tutor is a good example of misc 

where as the others were not...

and now nu-cli contains the cli stuff....

also I like the way you massaged the default_context...
nice effort on the cratification front...

we are slowly making progress...
I am kind of being patient (for the moment) as I see where things evolve with nu-std...

This will help define the cratification efforts as well...